+++ src/pages/Dashboard.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Reviews from '../components/Reviews'
+import Reviews from './Reviews'
 import CreateButton from '../components/CreateButton'
 
 export default function Dashboard({ reviews, onRemove, onEdit }) {
